+++ server/src/main/java/net/oneandone/stool/server/api/ApiController.java
@@ -362,18 +362,16 @@ public class ApiController {
     @GetMapping("/stages//{stage}/await-startup")
     public String awaitStartup(@PathVariable(value = "stage") String stageName) throws IOException {
         Stage stage;
-        JsonObject result;
 
         System.out.println("await new " + stageName);
         try (Engine engine = engine()) {
             stage = server.load(stageName);
             stage.awaitStartup(engine);
 
-            result = new JsonObject();
-            if (stage.currentOpt(engine) != null) {
-                result.add(Stage.APP_NAME, Engine.obj(stage.urlMap(engine, server.pool, Stage.APP_NAME)));
+            if (stage.currentOpt(engine) == null) {
+                throw new IllegalStateException();
             }
-            return result.toString();
+            return Engine.obj(stage.urlMap(engine, server.pool, Stage.APP_NAME)).toString();
         }
     }