+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   * `ssh` dumped app argument
   * `tunnel` dumped app argument
   * `build` dumped app result field
+  * `awaitStartup` now returns a single object reposentings the urls
   * naming scheme for context directories changed from app:tag to tag (current build is now _)
   * build now creates a repository tag without app name
   * dumped container label `app`